Refactor utility responsibilities
